requirement.txt
fastapi : 0.116.1
uvicorn : 0.35.0
pymongo : 4.13.2
python-dotenv : 1.1.1
passlib[bcrypt] : 1.7.4 
PyJWT : 2.10.1
pydantic[email] : 2.11.7
python-multipart :  0.0.20
python-jose[cryptography] : 3.5.0



